modify comparePackageOutput
