+++ chore/gogensig/cmptest/test.go
@@ -1,16 +1,13 @@
 package cmptest
 
 import (
-	"bufio"
 	"bytes"
 	"encoding/json"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
-	"unicode"
 
-	"github.com/google/go-cmp/cmp"
+	"github.com/goplus/llgo/chore/gogensig/cmp"
 	"github.com/goplus/llgo/chore/gogensig/unmarshal"
 	"github.com/goplus/llgo/chore/gogensig/util"
 	"github.com/goplus/llgo/chore/gogensig/visitor"
@@ -18,75 +15,6 @@ import (
 	"github.com/goplus/llgo/chore/gogensig/visitor/symb"
 	cppgtypes "github.com/goplus/llgo/chore/llcppg/types"
 )
-
-func skipSpace(data []byte, from int) int {
-	dataBuf := bytes.NewBuffer(data[from:])
-	index := from
-	for {
-		rn, sz, err := dataBuf.ReadRune()
-		if err != nil {
-			break
-		}
-		if !unicode.IsSpace(rn) {
-			break
-		}
-		index = index + sz
-	}
-	return index
-}
-
-func SplitLineIgnoreSpace(s string) []string {
-	buf := bytes.NewBufferString(s)
-	scan := bufio.NewScanner(buf)
-	results := make([]string, 0)
-	for scan.Scan() {
-		lineText := scan.Text()
-		lineTextBuf := bytes.NewBufferString(lineText)
-		lineTextScan := bufio.NewScanner(lineTextBuf)
-		lineTextScan.Split(func(data []byte, atEOF bool) (advance int, token []byte, err error) {
-			if atEOF && len(data) == 0 {
-				return 0, nil, nil
-			}
-			if i := bytes.IndexAny(data, " \t"); i >= 0 {
-				ii := skipSpace(data, i+1)
-				return ii, data[0:i], nil
-			}
-			if atEOF {
-				return len(data), data, nil
-			}
-			return 0, nil, nil
-		})
-
-		strBuilder := strings.Builder{}
-		writeSpace := false
-		total := 0
-		for lineTextScan.Scan() {
-			word := lineTextScan.Text()
-			if writeSpace {
-				strBuilder.WriteRune(' ')
-			}
-			n, err := strBuilder.WriteString(word)
-			if err != nil {
-				break
-			}
-			total += n
-			writeSpace = total > 0
-		}
-		if total > 0 {
-			results = append(results, strBuilder.String())
-		}
-	}
-	return results
-}
-
-func EqualStringIgnoreSpace(s1 string, s2 string) (bool, string) {
-	arr1 := SplitLineIgnoreSpace(s1)
-	arr2 := SplitLineIgnoreSpace(s2)
-	if !cmp.Equal(arr1, arr2) {
-		return false, cmp.Diff(arr1, arr2)
-	}
-	return true, ""
-}
 
 func RunTest(t *testing.T, pkgName string, isCpp bool, symbolEntries []symb.SymbolEntry, cppgConf *cppgtypes.Config, originalCode, expectedOutput string) bytes.Buffer {
 	t.Helper()
@@ -117,7 +45,7 @@ func RunTest(t *testing.T, pkgName string, isCpp bool, symbolEntries []symb.Symb
 	p.UnmarshalBytes(bytes)
 
 	result := buf.String()
-	if isEqual, diff := EqualStringIgnoreSpace(expectedOutput, result); !isEqual {
+	if isEqual, diff := cmp.EqualStringIgnoreSpace(expectedOutput, result); !isEqual {
 		t.Errorf("unexpected result:\n%s", diff)
 	}
